fix: Deleting contact used to show up in waContacts, add: forward messages barebones

+++ src/hooks/useSubscribeToUserContactsFromContactsApp.js
@@ -4,6 +4,7 @@ import {
   collection,
   deleteDoc,
   doc,
+  getDoc,
   onSnapshot,
   query,
   setDoc,
@@ -45,16 +46,27 @@ export function useSubscribeToUserContactsFromContactsApp(currentUserEmail) {
           if (change.type === 'removed') {
             // delete that contact from user's waContacts
             const encodedEmail = encodeEmail(contact.email);
-            setDoc(
+            getDoc(
               doc(db, 'whatsApp/userContacts', currentUserEmail, contact.email),
-              { [encodedEmail]: true, email: contact.email },
-            ).then(() => {
-              // delete the flag that started off this conversation so that it can be re-added by next incoming chat
-              const chatName = chatNameGenerator(
-                currentUserEmail,
-                contact.email,
-              );
-              deleteDoc(doc(db, 'whatsApp/chats', chatName, 'flag'));
+            ).then((doc) => {
+              if (doc.exists()) {
+                setDoc(
+                  doc(
+                    db,
+                    'whatsApp/userContacts',
+                    currentUserEmail,
+                    contact.email,
+                  ),
+                  { [encodedEmail]: true, email: contact.email },
+                ).then(() => {
+                  // delete the flag that started off this conversation so that it can be re-added by next incoming chat
+                  const chatName = chatNameGenerator(
+                    currentUserEmail,
+                    contact.email,
+                  );
+                  deleteDoc(doc(db, 'whatsApp/chats', chatName, 'flag'));
+                });
+              }
             });
           }
         });
